requirements.txt
alembic==1.7.4
Babel==2.9.1
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
dominate==2.6.0
elasticsearch==7.15.1
email-validator==1.1.3
Flask==2.0.2
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
langdetect==1.0.9
Mako==1.1.5
MarkupSafe==2.0.1
PyJWT==2.3.0
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.26
urllib3==1.26.7
visitor==0.1.3
Werkzeug==2.0.2
WTForms==2.3.3
